storiesのts化
