--- storybook/stories/0-Welcome.stories.js
@@ -1,17 +1,0 @@
-import { linkTo } from '@storybook/addon-links';
-
-import Welcome from './Welcome';
-
-export default {
-  title: 'Welcome',
-};
-
-export const toStorybook = () => ({
-  components: { Welcome },
-  template: '<welcome :showApp="action" />',
-  methods: { action: linkTo('Button') },
-});
-
-toStorybook.story = {
-  name: 'to Storybook',
-};
