--- storybook/stories/1-Button.stories.js
@@ -1,28 +1,0 @@
-import { action } from '@storybook/addon-actions';
-import { linkTo } from '@storybook/addon-links';
-
-import MyButton from './MyButton';
-
-export default {
-  title: 'Button',
-};
-
-export const text = () => ({
-  components: { MyButton },
-  template: '<my-button @click="action">Hello Button</my-button>',
-  methods: { action: action('clicked') },
-});
-
-export const jsx = () => ({
-  components: { MyButton },
-  render(h) {
-    return <my-button onClick={this.action}>With JSX</my-button>;
-  },
-  methods: { action: linkTo('clicked') },
-});
-
-export const emoji = () => ({
-  components: { MyButton },
-  template: '<my-button @click="action">😀 😎 👍 💯</my-button>',
-  methods: { action: action('clicked') },
-});
